refactor: move registry bootstrap out of RegistryLoaderMixin into OTGRegistryHelper
Mixin classes live under a classload-policed package and every javac
synthetic (enum-switch $SwitchMap, anonymous class) becomes a landmine
when the Forgix merge suffix-renames it (RegistryLoaderMixin$1_fabric
broke Outer$Inner companion detection and crashed world creation).
Move all logic into the regular class shared.registry.OTGRegistryHelper
(same FQCN as 1.21.1), restore the enum switches the hotfix flattened,
and drop the dead mixin.util.RegistryUtil (WorldPresetTagsMixin already
carries its own private getAsReference). The mixin package now holds
only thin @Mixin/accessor classes; fixMergedMixinConfigs enforces this
at build time.
